INDY-2306: tests to check that GET_CRED_DEF for large Schemas works
